Move login and signup to app router, refactor signup
